pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "reportlab>=4.4.2",
    "streamlit>=1.46.1",
]
